Create store for timing objects
