+++ electron/database/ApiUtils.ts
@@ -1,25 +1,28 @@
 import Beat from "../../src/global/classes/Beat";
 import Measure from "../../src/global/classes/Measure";
-import Page from "../../src/global/classes/Page";
 import { fromDatabaseBeat, getBeats } from "./tables/BeatTable";
 import Database from "better-sqlite3";
 import { fromDatabaseMeasures, getMeasures } from "./tables/MeasureTable";
 import { fromDatabasePages, getPages } from "./tables/PageTable";
+import { TimingObjects } from "../../src/stores/TimingObjectsStore";
+import { DatabaseResponse } from "./DatabaseActions";
 
-export const GenerateTimingObjects = ({
+export const generateTimingObjects = ({
     db,
 }: {
     db: Database.Database;
-}): {
-    beats: Beat[];
-    measures: Measure[];
-    pages: Page[];
-} => {
+}): DatabaseResponse<TimingObjects> => {
+    const failureObject: TimingObjects = {
+        beats: [],
+        measures: [],
+        pages: [],
+    };
+
     // Generate the beats
     const databaseBeatsResponse = getBeats({ db });
     if (!databaseBeatsResponse.success) {
         console.error(databaseBeatsResponse.error);
-        throw new Error("Failed to get beats from database");
+        return { ...databaseBeatsResponse, data: failureObject };
     }
     const beats: Beat[] = databaseBeatsResponse.data.map(fromDatabaseBeat);
 
@@ -27,7 +30,7 @@ export const GenerateTimingObjects = ({
     const databaseMeasuresResponse = getMeasures({ db });
     if (!databaseMeasuresResponse.success) {
         console.error(databaseMeasuresResponse.error);
-        throw new Error("Failed to get measures from database");
+        return { ...databaseBeatsResponse, data: failureObject };
     }
     const measures: Measure[] = fromDatabaseMeasures({
         databaseMeasures: databaseMeasuresResponse.data,
@@ -38,7 +41,7 @@ export const GenerateTimingObjects = ({
     const databasePagesResponse = getPages({ db });
     if (!databasePagesResponse.success) {
         console.error(databasePagesResponse.error);
-        throw new Error("Failed to get pages from database");
+        return { ...databaseBeatsResponse, data: failureObject };
     }
     const pages = fromDatabasePages({
         databasePages: databasePagesResponse.data,
@@ -47,8 +50,11 @@ export const GenerateTimingObjects = ({
     });
 
     return {
-        beats,
-        measures,
-        pages,
+        success: true,
+        data: {
+            beats,
+            measures,
+            pages,
+        },
     };
 };
