+++ electron/database/tables/PageTable.ts
@@ -6,9 +6,6 @@ import * as MarcherPageTable from "./MarcherPageTable";
 import { ModifiedMarcherPageArgs } from "../../../src/global/classes/MarcherPage";
 import { DatabaseBeat, getBeat } from "./BeatTable";
 import { getMarchers } from "./MarcherTable";
-import Beat from "../../../src/global/classes/Beat";
-import Measure from "../../../src/global/classes/Measure";
-import Page from "../../../src/global/classes/Page";
 
 export const FIRST_PAGE_ID = 0;
 
@@ -398,208 +395,3 @@ export function deletePages({
         data: response.data.map(realDatabasePageToDatabasePage),
     };
 }
-
-// Page Generation
-/** A type that stores a beat with the index that it occurs in a list with all beats */
-type BeatWithIndex = Beat & { index: number };
-
-/**
- * Converts the pages from the database (which are stored as a linked list) to Page objects.
- *
- * @param databasePages The pages from the database
- * @returns A list of Page objects
- */
-export function fromDatabasePages({
-    databasePages,
-    allMeasures,
-    allBeats,
-}: {
-    databasePages: DatabasePage[];
-    allMeasures: Measure[];
-    allBeats: Beat[];
-}): Page[] {
-    if (databasePages.length === 0) return [];
-    const sortedBeats = allBeats.sort((a, b) => a.position - b.position);
-    const beatMap = new Map<number, BeatWithIndex>(
-        sortedBeats.map((beat, i) => [beat.id, { ...beat, index: i }]),
-    );
-    const sortedDbPages = databasePages.sort((a, b) => {
-        const aBeat = beatMap.get(a.start_beat);
-        const bBeat = beatMap.get(b.start_beat);
-        if (!aBeat || !bBeat) {
-            console.log("aBeat", a.start_beat, aBeat);
-            console.log("bBeat", b.start_beat, bBeat);
-            throw new Error(
-                `Beat not found: ${a.start_beat} ${aBeat} - ${b.start_beat} ${bBeat}`,
-            );
-        }
-        return aBeat.position - bBeat.position;
-    });
-    const isSubsetArr = sortedDbPages.map((page) => page.is_subset);
-    const pageNames = generatePageNames(isSubsetArr);
-    const sortedMeasures = allMeasures.sort((a, b) => a.number - b.number);
-
-    let curTimestamp = 0;
-    const createdPages: Page[] = sortedDbPages.map((dbPage, i) => {
-        // Get the beats that belong to this page
-        const startBeat = beatMap.get(dbPage.start_beat);
-        if (!startBeat) {
-            throw new Error(`Start beat not found: ${dbPage.start_beat}`);
-        }
-        const nextPage = sortedDbPages[i + 1] || null;
-        const nextPageBeat = nextPage ? beatMap.get(nextPage.start_beat) : null;
-        if (!nextPageBeat && nextPage) {
-            throw new Error(`Next beat not found: ${nextPage.start_beat}`);
-        }
-        const beats = nextPage
-            ? sortedBeats.slice(startBeat.index, nextPageBeat!.index)
-            : sortedBeats.slice(startBeat.index);
-        const lastBeat = beats[beats.length - 1];
-        const beatIdSet = new Set(beats.map((beat) => beat.id));
-
-        // Get the measures that belong to this page
-        const measures = sortedMeasures.filter(
-            (measure) =>
-                // Check if the start beat of the measure is on or after the start beat of the page
-                (measure.startBeat.position >= startBeat.position ||
-                    // Check that the start beat is on or before the last beat of the page
-                    measure.startBeat.position <= lastBeat.position) &&
-                // If both are true, ensure that the beat is actually in the measure
-                measure.beats.some((beat) => beatIdSet.has(beat.id)),
-        );
-        if (measures.length === 0) {
-            throw new Error(`No measures found for page ${dbPage.id}`);
-        }
-        const duration = beats.reduce((acc, beat) => acc + beat.duration, 0);
-        const output = {
-            id: dbPage.id,
-            name: pageNames[i],
-            counts: beats.length,
-            notes: dbPage.notes || null,
-            order: i,
-            isSubset: dbPage.is_subset,
-            duration: duration,
-            beats,
-            measures,
-            measureBeatToStartOn:
-                measures[0].beats.findIndex(
-                    (beat) => beat.id === startBeat.id,
-                ) + 1,
-            measureBeatToEndOn:
-                measures[measures.length - 1].beats.findIndex(
-                    (beat) => beat.id === lastBeat.id,
-                ) + 1,
-            timestamp: curTimestamp,
-            previousPageId: sortedDbPages[i - 1]?.id || null,
-            nextPageId: nextPage ? nextPage.id : null,
-        } satisfies Page;
-        curTimestamp += duration;
-        return output;
-    });
-
-    return createdPages;
-}
-
-/**
- * Creates a list of page names based on the list of booleans that pages are subsets or not.
- *
- * E.g. [False, False, True, False, True, True, False] => ["1", "2", "2A", "3", "3A", "3B", "4"]
- *
- * NOTE - the first page will always evaluate to false no matter what is provided.
- *
- * @param pages boolean[] - A list to define if pages are subsets are not. Should align with the order of the pages.
- * @returns A list of page names in the order that the list of booleans was provided.
- */
-export const generatePageNames = (isSubsetArr: boolean[]) => {
-    const pageNames: string[] = ["0"];
-    let curPageNumber = 0;
-    let curSubsetLetter = "";
-
-    /**
-     * Increments a letter to the next letter in the alphabet.
-     *
-     * @param letters The letters to increment.
-     * @returns A -> B, B -> C, ..., Z -> AA, AA -> AB, etc. Letters are always capitalized.
-     */
-    const incrementLetters = (letters: string) => {
-        let result = [];
-        let carry = true; // Start with the assumption that we need to increment the last character
-
-        const capitalizedLetters = letters.toUpperCase();
-
-        // Traverse from last to first character to handle the carry
-        for (let i = capitalizedLetters.length - 1; i >= 0; i--) {
-            let char = capitalizedLetters[i];
-            if (carry) {
-                if (char === "Z") {
-                    result.push("A");
-                } else {
-                    result.push(String.fromCharCode(char.charCodeAt(0) + 1));
-                    carry = false; // No carry needed if we haven't wrapped from 'Z' to 'A'
-                }
-            } else {
-                result.push(char); // If no carry, keep current character as is
-            }
-        }
-
-        // If the string was all 'Z's, we will have carry left over after processing all characters
-        if (carry) {
-            result.push("A"); // Append 'A' to handle cases like 'ZZ' -> 'AAA'
-        }
-
-        // Since we've constructed the result in reverse order, reverse it back and join into a string
-        return result.reverse().join("");
-    };
-
-    /**
-     * Get the next page name based on the current page name.
-     *
-     * @param pageNumber The number of the current page.
-     * @param subsetString The subset letter of the current page. If null, the page is not a subset.
-     * @param incrementSubset Whether it is the number or the subset letter that should be incremented.
-     *      Default is false, which increments the number. If true, increments the subset letter and not the number.
-     * @returns
-     */
-    const getNextPageName = ({
-        pageNumber,
-        subsetString,
-        incrementSubset = false,
-    }: {
-        pageNumber: number;
-        subsetString: string | null;
-        incrementSubset?: boolean;
-    }) => {
-        let newPageNumber = pageNumber;
-        let newSubsetString = subsetString || "";
-
-        if (incrementSubset) {
-            // If there is no subset, start with "A"
-            if (!subsetString || subsetString === "") newSubsetString = "A";
-            // Otherwise, increment the subset letter
-            else newSubsetString = incrementLetters(subsetString);
-        } else {
-            newPageNumber = pageNumber + 1;
-            newSubsetString = "";
-        }
-
-        return newPageNumber + newSubsetString;
-    };
-
-    // Loop through the pages and create the page names
-    // 1, 2, 2A, 3, 3A, 3B, 4, etc.
-    for (let i = 1; i < isSubsetArr.length; i++) {
-        const pageName = getNextPageName({
-            pageNumber: curPageNumber,
-            subsetString: isSubsetArr[i] ? curSubsetLetter : null,
-            incrementSubset: isSubsetArr[i],
-        });
-        pageNames.push(pageName);
-        if (isSubsetArr[i]) {
-            curSubsetLetter = incrementLetters(curSubsetLetter);
-        } else {
-            curPageNumber++;
-            curSubsetLetter = "";
-        }
-    }
-    return pageNames;
-};
