+++ src/context/SelectedPageContext.tsx
@@ -5,8 +5,8 @@ import {
     useEffect,
     useState,
 } from "react";
-import { usePageStore } from "@/stores/PageStore";
 import Page from "@/global/classes/Page";
+import { useTimingObjectsStore } from "@/stores/TimingObjectsStore";
 
 // Define the type for the context value
 type SelectedPageContextProps = {
@@ -19,7 +19,7 @@ const SelectedPageContext = createContext<SelectedPageContextProps | undefined>(
 );
 
 export function SelectedPageProvider({ children }: { children: ReactNode }) {
-    const { pages } = usePageStore();
+    const { pages } = useTimingObjectsStore();
     const [selectedPage, setSelectedPage] = useState<Page | null>(null);
 
     // Send the selected page to the electron main process
