+++ src/global/classes/Beat.ts
@@ -1,3 +1,5 @@
+import { DatabaseBeat } from "electron/database/tables/BeatTable";
+
 /**
  * A Beat represents a specific point in time in the show.
  * It has a duration until the next beat and can be included in measures.
@@ -32,3 +34,18 @@ export const compareBeats = (a: Beat, b: Beat): number =>
  */
 export const beatsDuration = (beats: Beat[]): number =>
     beats.reduce((total, beat) => total + beat.duration, 0);
+
+/**
+ * Converts a DatabaseBeat object to a Beat object.
+ * @param beat - The DatabaseBeat object to convert.
+ * @returns A new Beat object with the same properties as the input DatabaseBeat.
+ */
+export const fromDatabaseBeat = (beat: DatabaseBeat): Beat => {
+    return {
+        id: beat.id,
+        position: beat.position,
+        duration: beat.duration,
+        includeInMeasure: beat.include_in_measure >= 1,
+        notes: beat.notes,
+    };
+};
