+++ src/stores/TimingObjectsStore.ts
@@ -1,0 +1,70 @@
+import Beat, { fromDatabaseBeat } from "@/global/classes/Beat";
+import Measure, { fromDatabaseMeasures } from "@/global/classes/Measure";
+import Page, { fromDatabasePages } from "@/global/classes/Page";
+import { create } from "zustand";
+
+export type TimingObjects = {
+    beats: Beat[];
+    measures: Measure[];
+    pages: Page[];
+};
+
+type TimingObjectStoreInterface = TimingObjects & {
+    /**
+     * Fetch the timing objects (beats, measures, and pages) from the database and updates the store.
+     * @returns A promise that resolves when the TimingObjects have been fetched
+     */
+    fetchTimingObjects: () => Promise<void>;
+};
+
+export const useTimingObjectsStore = create<TimingObjectStoreInterface>(
+    (set) => ({
+        pages: [],
+        measures: [],
+        beats: [],
+
+        /**
+         * Fetch the pages from the database and updates the store.
+         * This is the only way to update retrieve the pages from the database that ensures the UI is updated.
+         */
+        fetchTimingObjects: async (): Promise<void> => {
+            const pagesResponse = await window.electron.getPages();
+            if (!pagesResponse.success) {
+                console.error(pagesResponse.error);
+                throw new Error(
+                    pagesResponse.error?.message ?? "Could not fetch pages",
+                );
+            }
+            const beatsResponse = await window.electron.getBeats();
+            if (!beatsResponse.success) {
+                console.error(beatsResponse.error);
+                throw new Error(
+                    beatsResponse.error?.message ?? "Could not fetch beats",
+                );
+            }
+            const measuresResponse = await window.electron.getMeasures();
+            if (!measuresResponse.success) {
+                console.error(measuresResponse.error);
+                throw new Error(
+                    measuresResponse.error?.message ??
+                        "Could not fetch measures",
+                );
+            }
+            const createdBeats = beatsResponse.data.map(fromDatabaseBeat);
+            const createdMeasures = fromDatabaseMeasures({
+                databaseMeasures: measuresResponse.data,
+                allBeats: createdBeats,
+            });
+            const createdPages = fromDatabasePages({
+                databasePages: pagesResponse.data,
+                allMeasures: createdMeasures,
+                allBeats: createdBeats,
+            });
+            set({
+                pages: createdPages,
+                measures: createdMeasures,
+                beats: createdBeats,
+            });
+        },
+    }),
+);
